requirements.txt
channel3-sdk>=3.0.0
anthropic>=0.88.0
fastapi>=0.128.0
uvicorn[standard]>=0.39.0
pydantic>=2.12.0
httpx>=0.28.0
python-dotenv>=1.2.0
